[OpenGL] update opengl

+++ Framework/Common/Allocator.hpp
@@ -51,12 +51,12 @@ namespace My {
         size_t m_szPageSize;
         size_t m_szAlignmentSize;
         size_t m_szBlockSize;
-        uint32_t m_nBlocksPerPage;
+        size_t m_nBlocksPerPage;
 
         //statistics 
-        uint32_t m_nPages;
-        uint32_t m_nBlocks;
-        uint32_t m_nFreeBlocks;
+        size_t m_nPages;
+        size_t m_nBlocks;
+        size_t m_nFreeBlocks;
 
         //disable copy&assignment
         Allocator(const Allocator& clone);
